Refactor file paths and imports for consistency
