+++ src/pages/Bandeau.js
@@ -1,4 +1,4 @@
-import '../styles_composants/Bandeau.css';
+import '../css/Bandeau.css';
 import ElemBandeau from './ElemBandeau';
 import ElemReseau from './ElemReseau';
